Move prune_unused_parameters to OIR pass
