+++ src/gtc/passes/gtir_pipeline.py
@@ -19,7 +19,6 @@ from typing import Callable, Dict, Optional, Sequence, Tuple
 from gtc import gtir
 from gtc.passes.gtir_definitive_assignment_analysis import check as check_assignments
 from gtc.passes.gtir_dtype_resolver import resolve_dtype
-from gtc.passes.gtir_prune_unused_parameters import prune_unused_parameters
 from gtc.passes.gtir_upcaster import upcast
 
 
@@ -38,7 +37,7 @@ class GtirPipeline:
         self._cache: Dict[Tuple[PASS_T, ...], gtir.Stencil] = {}
 
     def steps(self) -> Sequence[PASS_T]:
-        return [check_assignments, prune_unused_parameters, resolve_dtype, upcast]
+        return [check_assignments, resolve_dtype, upcast]
 
     def apply(self, steps: Sequence[PASS_T]) -> gtir.Stencil:
         result = self.gtir
