+++ src/gtc/passes/oir_pipeline.py
@@ -31,7 +31,11 @@ from gtc.passes.oir_optimizations.horizontal_execution_merging import (
 )
 from gtc.passes.oir_optimizations.inlining import MaskInlining
 from gtc.passes.oir_optimizations.mask_stmt_merging import MaskStmtMerging
-from gtc.passes.oir_optimizations.pruning import NoFieldAccessPruning, UnreachableStmtPruning
+from gtc.passes.oir_optimizations.pruning import (
+    NoFieldAccessPruning,
+    UnreachableStmtPruning,
+    prune_unused_parameters,
+)
 from gtc.passes.oir_optimizations.temporaries import (
     LocalTemporariesToScalars,
     WriteBeforeReadTemporariesToScalars,
@@ -69,6 +73,7 @@ class DefaultPipeline(OirPipeline):
     @staticmethod
     def all_steps() -> Sequence[PassT]:
         return [
+            prune_unused_parameters,
             AdjacentLoopMerging,
             HorizontalExecutionMerging,
             OnTheFlyMerging,
